+++ Assets/SteamMultiRuntime/Runtime/Packages/com.koiusa.keyconfig/Editor/AssemblyInfo.cs
@@ -1,3 +1,0 @@
-using System.Runtime.CompilerServices;
-
-[assembly: InternalsVisibleTo("Koiusa.KeyConfig.Editor.Tests")]